+++ cmd/samples/pageflow/src/server/helpers/index.js
@@ -1,9 +1,10 @@
 export { default as createProduct } from './createProduct.js';
 export { default as getProduct } from './getProduct.js';
+export { default as getProductOrRetry } from './getProductOrRetry.js';
 export { default as handleError } from './handleError.js';
 export { default as momentToLong } from './momentToLong.js';
 export { default as parseContent } from './parseContent.js';
-export { default as signalPageWorkflow } from './signalPageWorkflow.js';
+export { default as signalAndGetProduct } from './signalAndGetProduct.js';
 export { default as startPageFlowWorkflow } from './startPageFlowWorkflow.js';
 export { default as updateProductDescription } from './updateProductDescription.js';
 export { default as updateProductState } from './updateProductState.js';
