+++ cmd/samples/pageflow/src/server/helpers/updateProductState.js
@@ -1,57 +1,22 @@
 import {
-  EXPECTED_STATUS_FROM_STATE,
   ALLOWED_STATE_ON_STATUS_MAP,
 } from '../constants.js';
-import config from '../config.js';
 import getProduct from './getProduct.js';
-import signalPageWorkflow from './signalPageWorkflow.js';
-import waitTime from './waitTime.js';
-
-const { retryDelay, retryMax } = config.cadence;
-
-const getProductOrRetry = async ({
-  cadence,
-  expectedStatus,
-  name,
-  retry = 0,
-  status = ''
-}) => {
-  if (retry === retryMax) {
-    throw new UnexpectedStatusError(status);
-  }
-
-  const product = await getProduct({ cadence, name });
-
-  if (product.status !== expectedStatus) {
-    await waitTime(retryDelay);
-    return getProductOrRetry({
-      cadence,
-      expectedStatus,
-      name,
-      retry: retry + 1,
-      status: product.status,
-    });
-  }
-
-  return product;
-};
+import signalAndGetProduct from './signalAndGetProduct.js';
 
 const updateProductState = async ({ cadence, name, state }) => {
   const product = await getProduct({ cadence, name });
-  const expectedStatus = EXPECTED_STATUS_FROM_STATE[state];
 
   if (!ALLOWED_STATE_ON_STATUS_MAP[state].includes(product.status)) {
     throw new UnexpectedStatusError(product.status);
   }
 
-  await signalPageWorkflow({
+  return signalAndGetProduct({
     cadence,
     content: '',
     name,
     state,
   });
-
-  return getProductOrRetry({ cadence, expectedStatus, name });
 };
 
 export default updateProductState;
